refactor: collapse todo handlers into a single resource handler
