+++ test.sh
@@ -10,8 +10,8 @@ CREATE_TODO_REQUEST='{"body": "foo"}'
 UPDATE_TODO_REQUEST='{"body": "foobar", "completed": true}'
 
 set -x
-# test user create
-curl -i -X POST ${API_HOSTNAME}/${USER_ENDPOINT} -d "${CREATE_USER_REQUEST}"
+# # test user create
+# curl -i -X POST ${API_HOSTNAME}/${USER_ENDPOINT} -d "${CREATE_USER_REQUEST}"
 
 # test todo create
 resp=`curl -X POST ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT} -d "${CREATE_TODO_REQUEST}"`
@@ -24,10 +24,10 @@ curl -i ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}
 # test todo get-one
 curl -i ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId}
 
-# # test todo delete-one
-# curl -i -X DELETE ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId}
-# curl -i ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId}
-
 # test todo update-one
 curl -i -X PUT ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId} -d "${UPDATE_TODO_REQUEST}"
+curl -i ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId}
+
+# test todo delete-one
+curl -i -X DELETE ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId}
 curl -i ${API_HOSTNAME}/${USER_ENDPOINT}/shawn/${TODOS_ENDPOINT}/${todoId}
